solve day 9 part 1

+++ src/day09/Day09.kt
@@ -31,7 +31,7 @@ fun main() {
         }
     }
 
-    fun part1(input: List<String>): Int {
+    fun part1(input: List<String>): Long {
         val diskMap = parseDiskMap(input).toMutableList()
 
         while (!diskMap.isCompact()) {
@@ -39,16 +39,16 @@ fun main() {
         }
 
         return diskMap.filterIsInstance<DiskMapEntry.File>().mapIndexed { index, diskMapEntry ->
-            index * diskMapEntry.id
+            index.toLong() * diskMapEntry.id.toLong()
         }.sum()
     }
 
-    fun part2(input: List<String>): Int {
-        return -1
+    fun part2(input: List<String>): Long {
+        return -1L
     }
 
     val testInput = readInput("Day09_test")
-    check(part1(testInput) == 1928)
+    check(part1(testInput) == 1928L)
 //    check(part2(testInput) == 0)
 
     val input = readInput("Day09")
